feat: check if spot empty, and board full
We will need to check if a spot is full in order to validate if a user
can place their token there or not. When the user chooses a column to
place their token into, we will call empty? on the elements to make
sure there is at least 1 element free at the top.

Also, for checking if the game is a draw, we will need to see if the
"board is full" AND "there is no winner", so the full? method on the
board will help with this later on.

This was also done using TDD, so the tests for each function are
included.