feat: check if board has a winner
In order to terminate the game, one of the conditions to check is if
someone has won the game. You can win the game in one of 4 ways:
1. Have four of the same colour in a row
2. Have four of the same colour in a column
3. Have four of the same colour in a high-to-low diagonal
4. Have four of the same colour in a low-to-high diagonal

The winner? method when called will iterate through all of the above
cases across the entire board, and check if someone has won the game
or not.

This was done using TDD so there are also tests written for it.